feature(gameOver): Color constants file
