+++ WavePong/Fonts/ColorConstants.swift
@@ -1,0 +1,17 @@
+//
+//  ColorConstants.swift
+//  WavePong
+//
+//  Created by Lucas Migge on 20/08/23.
+//
+
+import Foundation
+
+/// Object of acessing custom color names
+struct ColorConstants {
+    static var AMARELO = "amarelo"
+    static var LARANJA = "laranja"
+    static var WHITE = "WHITE"
+    static var YELLOW = "Yellow"
+    
+}
