+++ WavePong/Fonts/Fonts Extension.swift
@@ -57,11 +57,3 @@ extension Font {
         
     }
 }
-
-struct ColorConstants {
-    static var AMARELO = "amarelo"
-    static var LARANJA = "laranja"
-    static var WHITE = "WHITE"
-    static var YELLOW = "Yellow"
-    
-}
